% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{main function to run linear regression with mini bag bootstrap}
\usage{
blblm(formula, data, m = 10, B = 5000, cluster_size = 1)
}
\arguments{
\item{formula}{an object with class "formula"}

\item{data}{a data frame or a vector of filenames}

\item{m}{how many times we split the data}

\item{B}{number of bootstrap performed}

\item{cluster_size}{number of clusters}
}
\value{
blblm object
}
\description{
main function to run linear regression with mini bag bootstrap
}
