% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{obtain confidence interval}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{blblm object}

\item{parm}{parms}

\item{level}{1-alpha, confidence level}

\item{...}{more argument}
}
\description{
obtain confidence interval
}
