% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm1}
\alias{lm1}
\title{run regression model}
\usage{
lm1(formula, data, freqs)
}
\arguments{
\item{formula}{a formula class}

\item{data}{a data frame}

\item{freqs}{number of times each rows are reselected in bootstrap}
}
\description{
run regression model
}
