% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_boot}
\alias{lm_each_boot}
\title{linear regression for each bootstrap}
\usage{
lm_each_boot(formula, data, n)
}
\arguments{
\item{formula}{a formula class}

\item{data}{a data frame}

\item{n}{sample size}
}
\description{
linear regression for each bootstrap
}
