% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_subsample}
\alias{lm_each_subsample}
\title{linear regression on each subsample}
\usage{
lm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{a formula class object}

\item{data}{a data frame}

\item{n}{sample size}

\item{B}{number of bootstrap}
}
\description{
linear regression on each subsample
}
