% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{prediciton for blbm class}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{blblm object}

\item{new_data}{a new data frame}

\item{confidence}{need confidence interval or not}

\item{level}{1-alpha, confidence level}

\item{...}{more argument}
}
\description{
prediciton for blbm class
}
