% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{obtain sigma}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{blblm class}

\item{confidence}{need confidence interval or not}

\item{level}{1-alpha, confidence level}

\item{...}{more arguments}
}
\description{
obtain sigma
}
