% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{split_data}
\alias{split_data}
\title{function to split data to m almost equal size data}
\usage{
split_data(data, m)
}
\arguments{
\item{data}{a data frame}

\item{m}{number of times to split data}
}
\description{
function to split data to m almost equal size data
}
